INTYGFV-14572: Merged master into branch.
